Create Template Class
create template class for to create new objects ar.
